+++ packages/tiny-canvas/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @dfosco/tiny-canvas
 
+## 3.10.0-beta.1
+
+## 3.10.0-beta.0
+
 ## 3.9.1
 
 ### Patch Changes